get rid of warnings

+++ src/neopixel.cpp
@@ -16,7 +16,7 @@ void NeoPixel::init() {
     );
 
     uint16_t buffer_length = 3 * strand_length;
-    uint8_t buf_length_data[] = {(buffer_length >> 8), buffer_length};
+    uint8_t buf_length_data[] = {(uint8_t)(buffer_length >> 8), (uint8_t)buffer_length};
 
     seesaw.write(
         MODULE_BASE,
@@ -30,7 +30,7 @@ void NeoPixel::init() {
 void NeoPixel::set(uint16_t number, uint8_t r, uint8_t g, uint8_t b) {
     number *= 3; // each pixel is 3 bytes in the buffer
 
-    uint8_t data[] = {(number >> 8), (number & 0xff), g, r, b};
+    uint8_t data[] = {(uint8_t)(number >> 8), (uint8_t)(number & 0xff), g, r, b};
 
     seesaw.write(
         MODULE_BASE,
